+++ c-dependencies/js-compute-runtime/builtins/logger.h
